Themes: Remove redundant actions and reducers
It is not necessary to have both THEMES_RECEIVE and THEMES_REQUEST_SUCCESS
action types, since they both signify an array of themes being returned
from a server.

Remove THEMES_RECEIVE and allow the receiveThemes() action to accept
a query, allowing it to take the place of receiveThemesQuery().

Inline the 'filterThemes' function to simplify found-count handling.